When broadcasting on iOS 16.4, the bitrate will become very low and result in a low-quality stream
If Liver is broadcasting a livestream using an iOS 16.4 device, the bitrate may drop to around 200k (normal range is 800k-1200k), resulting in a very low-quality stream.

Currently, this issue only occurs in iOS 16.4. It is possible that Apple will address this issue in a future iOS version, but it should be resolved in iOS 16.4.

Therefore, do not use DataRateLimit setting now.